Moved post_delete signal to s3file.py
